refactor: relocate several utilities to libs directory
