+++ src/components/settings/context/settings-provider.tsx
@@ -3,7 +3,7 @@ import { useEffect, useMemo, useCallback, useState } from 'react';
 // hooks
 import { useLocalStorage } from '@src/hooks/use-local-storage';
 // utils
-import { localStorageGetItem } from '@src/utils/storage-available';
+import { localStorageGetItem } from '@src/libs/storage-available';
 //
 import { SettingsValueProps } from '../types';
 import { SettingsContext } from './settings-context';
