+++ src/hooks/use-referrals.ts
@@ -10,7 +10,7 @@ import {
   checkIfEmailAlreadyAccepted as checkIfEmailAlreadyAcceptedAction,
   sendInvitation as sendInvitationAction,
   acceptOrCreateInvitationForUser as acceptOrCreateInvitationForUserAction,
-} from '@src/utils/supabase-actions';
+} from '@src/libs/supabase-actions';
 import { GLOBAL_CONSTANTS } from '@src/config-global';
 
 /**
