+++ src/libs/components/carousel/types.ts
