+++ src/locales/i18n.ts
@@ -2,7 +2,7 @@ import i18n from 'i18next';
 import LanguageDetector from 'i18next-browser-languagedetector';
 import { initReactI18next } from 'react-i18next';
 // utils
-import { localStorageGetItem } from '@src/utils/storage-available';
+import { localStorageGetItem } from '@src/libs/storage-available';
 //
 import { defaultLang } from './config-lang';
 //
